% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_flagged_dataset.R
\name{create_flagged_dataset}
\alias{create_flagged_dataset}
\title{This take a dataset generated by get_weather_data (or any of the snowload2::get_x_data functions),
and flaggs points which were found to be outliers during the national snow load project in 2020.}
\usage{
create_flagged_dataset(data, date_max = base::as.Date("2020-8-1"))
}
\arguments{
\item{data}{the dataset which is to be flagged}

\item{date_max}{observations after this as.Date() will be removed}
}
\value{
data frame which has
}
\description{
Since outliers were only flagged up to the date 2020-8-1, the default date_max is set to this.
This way observations after this point wont be marked as non-outliers when they haven't been checked.
It is advised to leave this date as it is.
}
\details{
0 means not an outlier, 1 means outlier
}
\examples{
# UT <- get_weather_data("UT")
# UT_flagged <- create_flagged_dataset(UT)
}
