% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_variables.R
\name{create_model_variables}
\alias{create_model_variables}
\title{Takes a flagged dataset created by create_flagged_dataset()}
\usage{
create_model_variables(data, complete = FALSE)
}
\arguments{
\item{data}{a flagged dataset}

\item{complete}{if true, uses complete.cases}
}
\value{
dataset with extra variables
}
\description{
Removes marked WESD values (as they are the only other variable we checked for)
OUTLIER_FINAL is if the SNWD was an outlier
Adds WESD, SNWD, PRCP, SNOW, TMIN, TMAX variables
}
\examples{
# UT <- create_flagged_dataset(get_weather_data("UT"))
# UT_model_variables <- create_model_variables(UT)
}
