% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcnd_monthly.R
\docType{data}
\name{ghcnd_monthly}
\alias{ghcnd_monthly}
\title{This data set is a summary of the GHCN-D on the monthly scale.
Stations/months with less than 5 observations were removed,
or if PRISM variables could not be matched to the station.
Only includes SNWD observations.}
\format{
A data frame with 671352 rows and 9 variables:
\describe{
\item{ID}{The weather station ID.}
\item{YEAR}{The year of the corresponding month.}
\item{MONTH}{The month of the observation.}
\item{MEAN}{The arithmetic average of the SNWD observations for the month.}
\item{MEDIAN}{The median of the SNWD observations.}
\item{SD}{The standard deviation of the observations.}
\item{MAX}{The maximum SNWD observation for the month.}
\item{MIN}{The minimum SNWD observation for the month.}
\item{RANGE}{The maximum SNWD observation minus the smallest SNWD observation.}
\item{Q1}{The 25th percentile of the months SNWD observations.}
\item{Q3}{The 75th percentile of the months SNWD observations.}
\item{N}{The number of observations (days) in the month.}
\item{MAXDIFFSCALED}{The maximum of the difference between observations when divided by the number of days between observations.}
\item{MINDIFFSCALED}{The minimum of the difference between observations when divided by the number of days between observations.}
\item{MAXDIFFRAW}{The maximum difference between consecutive observations (not necessarily 1 day apart), approximately the largest snowfall event for the month.}
\item{MINDIFFRAW}{The minimum difference between consecutive observations (not necessarily 1 day apart), approximately the largest melting of snow for the month.}
\item{PRISM_PPT}{The stations monthly precipitation as determined by PRISM climate normals.}
\item{PRISM_TMIN}{The stations monthly minimum temperature as determined by PRISM climate normals.}
\item{PRISM_TMAX}{The stations monthly maximum temperature as determined by PRISM climate normals.}
}
}
\source{
snowload2::outliers_combined modified by Brennan Bean to add the name
}
\usage{
ghcnd_monthly
}
\description{
This data set is a summary of the GHCN-D on the monthly scale.
Stations/months with less than 5 observations were removed,
or if PRISM variables could not be matched to the station.
Only includes SNWD observations.
}
\keyword{datasets}
