% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_ghcnd.R
\docType{data}
\name{outliers_ghcnd}
\alias{outliers_ghcnd}
\title{A data set which is formated for use in the create_flagged_Dataset function}
\format{
A data frame with 671352 rows and 9 variables:
\describe{
\item{ID}{The weather station ID}
\item{DATE}{Date of the flagged outlier}
\item{ELEMENT}{The element type of the outlier, i.e. SNWD or WESD}
\item{VALUE}{The value in mm (WESD will be in 10ths of mm, divide by 10 to get mm)}
\item{FLAGGED}{The method used to flag the outlier}
\item{OUTLIER}{Will always be 1 since this contains all outliers}
\item{TYPE}{Further describes how the observation was flagged}
\item{YEAR}{The year of the outlier}
\item{name}{The name of who flagged the outlier}
}
}
\source{
snowload2::outliers_combined modified by Brennan Bean to add the name
}
\usage{
outliers_ghcnd
}
\description{
It is a modified version of snowload2::outliers_combined
}
\keyword{datasets}
