% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM_climate_norms.R
\docType{data}
\name{PRISM_climate_norms}
\alias{PRISM_climate_norms}
\title{This rasterbrick has precipitation, min temp, and max temp for all 12 months.
Its primary use is in the create_model_variables() function}
\format{
A data frame with 36 layers, 12 for each month, with precipitation, min temp, and max temp,
and a final variable with the stations IDs
}
\source{
\url{https://prism.oregonstate.edu/normals/}
}
\usage{
PRISM_climate_norms
}
\description{
This rasterbrick has precipitation, min temp, and max temp for all 12 months.
Its primary use is in the create_model_variables() function
}
\keyword{datasets}
