% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_dataset.R
\name{subset_dataset}
\alias{subset_dataset}
\title{Subset_dataset}
\usage{
subset_dataset(data, zeros_to_one = 3, seed = 42069)
}
\arguments{
\item{data}{from create_model_variables}

\item{zeros_to_one}{number of 0's to 1's, default is 3 0's to 1 1}

\item{seed}{number to set seed so randomization is consistent}
}
\value{
dataframe
}
\description{
After creating a flagged dataset, and creating the model variables,
use this function to create training datasets.
}
\details{
Since there are significantly more non-outliers to outliers (zeros to ones),
it is nessicary to under-sample the zeros while training models.

The default is zeros_to_one=3, meaning there are 3 '0' observations for every
'1' observation. Every '1' observation is kept, and the '0's are randomly
sampled. The results are also shuffled.
}
\examples{
# UT <- get_weather_data("UT", "D:/Data/ghcnd_all/")
# data <- create_flagged_dataset(UT)
# data_1 <- create_model_variables(data)

# subset <- subset_dataset(data_1)

# trainSize <- .5
# train <- subset[1:(trainSize*nrow(subset)), ]
# test <- subset[(trainSize*nrow(subset)):nrow(subset), ]

}
