% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Title}
\usage{
plot_outliers(data, id)
}
\arguments{
\item{data}{full data frame with at least SNWD, OUTLIER_FINAL, ID, DATE, OUTLIER_PRED}

\item{id}{the id for the station you want to plot}
}
\value{
plotly plot
}
\description{
Title
}
